requirements.txt
pandas>=2.3.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
jupyter>=1.0.0
dvc>=3.0.0
xgboost>=2.0.0
shap>=0.44.0
scipy>=1.11.0
statsmodels>=0.14.0
pyyaml>=6.0
jupyterlab>=4.0.0
notebook>=7.0.0